social_media/requirements.txt
# django
django==4.1.1
python-dotenv==0.21.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
django-filter==23.2

# db-tools
psycopg2-binary==2.8.4